Break up App.js into focused hooks and layout components
App.js reduced from 1083 to 194 lines. Business logic extracted into
useCharacterStore, useCharacterHandlers, useAbilityPayment,
useRoundActions, and useCharacterManager. JSX split into MobileLayout
and DesktopLayout components. No behavior changed.

Also adds .tool-versions for asdf Node management, rewrites README,
and adds docs/ARCHITECTURE.md, docs/ONBOARDING.md, docs/ROADMAP.md.
