+++ .tool-versions
@@ -1,0 +1,1 @@
+nodejs 20.19.2
