+++ README.md
@@ -1,70 +1,42 @@
-# Getting Started with Create React App
+# Sturdy Pancake — RPG Character Sheet
 
-This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
+A web-based character sheet for a custom tabletop RPG system. Manage multiple characters, track attributes, skills, vitals, abilities, and combat techniques — all saved locally in your browser and shareable via URL.
 
-## Available Scripts
+**Live app:** https://paul-utley.github.io/sturdy-pancake-react-character-sheet/
 
-In the project directory, you can run:
+## Features
 
-### `npm start`
+- Multiple character support with quick switching
+- Attributes, skills, vitals (Health / Vigor / Dashes), and momentum tracking
+- Ability management with automatic cost deduction
+- Combat technique builder with point-cost calculation and trait selection
+- Drag-and-drop reordering for abilities and combat technique traits
+- Lock mode to prevent accidental edits during play
+- New Round / Quick Rest / Long Rest actions with field highlighting
+- Character sharing via compressed URL
+- Import / export characters as JSON
 
-Runs the app in the development mode.\
-Open [http://localhost:3000](http://localhost:3000) to view it in your browser.
+## Quick Start
 
-The page will reload when you make changes.\
-You may also see any lint errors in the console.
+Requires Node 20 (see `.tool-versions` — use [asdf](https://asdf-vm.com) to manage).
 
-### `npm test`
+```bash
+asdf install        # installs Node version from .tool-versions
+npm install
+npm start           # opens http://localhost:3000
+```
 
-Launches the test runner in the interactive watch mode.\
-See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
+## Scripts
 
-### `npm run build`
+| Command | Description |
+|---|---|
+| `npm start` | Start development server |
+| `npm test` | Run tests in watch mode |
+| `npm run build` | Production build |
+| `npm run deploy` | Build and deploy to GitHub Pages |
 
-Builds the app for production to the `build` folder.\
-It correctly bundles React in production mode and optimizes the build for the best performance.
+## Documentation
 
-The build is minified and the filenames include the hashes.\
-Your app is ready to be deployed!
-
-See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
-
-### `npm run eject`
-
-**Note: this is a one-way operation. Once you `eject`, you can't go back!**
-
-If you aren't satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
-
-Instead, it will copy all the configuration files and the transitive dependencies (webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you're on your own.
-
-You don't have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn't feel obligated to use this feature. However we understand that this tool wouldn't be useful if you couldn't customize it when you are ready for it.
-
-## Learn More
-
-You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
-
-To learn React, check out the [React documentation](https://reactjs.org/).
-
-### Code Splitting
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/code-splitting](https://facebook.github.io/create-react-app/docs/code-splitting)
-
-### Analyzing the Bundle Size
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size](https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size)
-
-### Making a Progressive Web App
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app](https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app)
-
-### Advanced Configuration
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/advanced-configuration](https://facebook.github.io/create-react-app/docs/advanced-configuration)
-
-### Deployment
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/deployment](https://facebook.github.io/create-react-app/docs/deployment)
-
-### `npm run build` fails to minify
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify](https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify)
+- [docs/ARCHITECTURE.md](docs/ARCHITECTURE.md) — tech stack, project structure, state management, data shapes
+- [docs/ONBOARDING.md](docs/ONBOARDING.md) — getting started guide for new contributors
+- [docs/ROADMAP.md](docs/ROADMAP.md) — planned features and known code quality work
